LayoutTests/platform/ios-simulator/editing/deleting/delete-and-cleanup-expected.txt

This test checks that deletion does not leave unnecessary nested divs.

FAIL confirmedMarkup should be <br>. Was Hello.
FAIL confirmedMarkup should be <br>. Was <div>Hello</div>.
FAIL confirmedMarkup should be <div id="mydiv"><br></div>. Was <div id="mydiv">Hello</div>.
FAIL confirmedMarkup should be <br>. Was <div><div>Hello</div></div>.
FAIL confirmedMarkup should be <b><br></b>. Was <div><b><div><i>Hello</i></div></b></div>.
FAIL confirmedMarkup should be <b><div style="border: solid red"><br></div></b>. Was <div><b><div style="border: solid red"><i>Hello</i></div></b></div>.
FAIL confirmedMarkup should be <br>. Was <div>Hello</div><div>world</div>.
FAIL confirmedMarkup should be <br>. Was <div>Hello</div><div>world</div>.
FAIL confirmedMarkup should be <p><br></p>. Was <div><p>Hello</p></div><p>world</p>.
PASS successfullyParsed is true

TEST COMPLETE
